third_party/WebKit/LayoutTests/fast/dom/custom/unresolved-pseudoclass-expected.txt

Tests the :unresolved pseudoclass.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.querySelector("x-x:unresolved") is a
PASS window.getComputedStyle(a).color is "rgb(0, 0, 222)"
PASS window.getComputedStyle(b).color is "rgb(0, 0, 222)"
PASS window.getComputedStyle(b).borderColor is "rgb(0, 0, 111)"
PASS window.getComputedStyle(c).color is "rgb(0, 222, 0)"
PASS window.getComputedStyle(a).color is "rgb(0, 222, 0)"
PASS window.getComputedStyle(d).color is "rgb(0, 111, 0)"
PASS window.getComputedStyle(b).color is "rgb(0, 111, 0)"
PASS successfullyParsed is true

TEST COMPLETE
